Split leaderboard view into multiple components
